% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brar_select_au_binary.r
\name{brar_select_au_binary}
\alias{brar_select_au_binary}
\title{brar_select_au_binary}
\usage{
brar_select_au_binary(
  Pats,
  nMax,
  TimeToOutcome,
  enrollrate,
  N1,
  armn,
  h,
  N2,
  tp,
  armlabel,
  blocksize,
  alpha1 = 1,
  beta1 = 1,
  alpha2 = alpha1,
  beta2 = beta1,
  alpha3 = alpha1,
  beta3 = beta1,
  alpha4 = alpha1,
  beta4 = beta1,
  alpha5 = alpha1,
  beta5 = beta1,
  minstart,
  deltaa,
  tpp = 0,
  deltaa1,
  side,
  output = NULL,
  ...
)
}
\arguments{
\item{Pats}{the number of patients accrued within a certain time frame indicates the
count of individuals who have been affected by the disease during that specific period,
for example, a month or a day. If this number is 10, it represents that
10 people have got the disease within the specified time frame.}

\item{nMax}{the maximal accrued number of patients with the disease, this number
should be chosen carefully to ensure a sufficient number of patients are simulated,
especially when considering the delay mechanism.}

\item{TimeToOutcome}{Representation of the time distribution of delayed responses. The accrual times
could be a month, a week or any other time frame. When the unit changes,
the number of TimeToOutcome should also change. It can be in the format
of expression(rnorm( length( vStartTime ),30, 3)), representing delayed responses
with a normal distribution, where the mean is 30 days and the standard deviation is 3 days.
These related functions are adapted from \url{http://github.com/kwathen/IntroBayesianSimulation}.
Refer to the website for more details.}

\item{enrollrate}{probability that patients in the population can enroll in the trial.
This parameter is related to the number of people who have been affected by the disease in the population,
following an exponential distribution.}

\item{N1}{number of participants with equal randomization in the ''initialization''initialization' period.
Recommend using 10 percent of the total sample size.}

\item{armn}{number of total arms in the trial.}

\item{h}{a vector of success probabilities in hypotheses, for example, as c(0.1,0.1) where 0.1 stands for the success probability
for both groups. Another example is c(0.1,0.3) where 0.1 and 0.3 stand for the success probabilities
for the control and the treatment group, respectively.}

\item{N2}{maximal sample size for the trial.}

\item{tp}{tuning parameter. Some commonly used numbers are 0.5, 1 and n/2N.}

\item{armlabel}{a vector of arm labels with an example of c(1, 2), where 1 and 2 describe
how each arm is labeled in a two-armed trial.}

\item{blocksize}{size of block used for equal randomization regarding participants in the 'initialization' period.
Recommend to be an even multiple of the number of total arms.}

\item{alpha1}{\eqn{\alpha} in the \eqn{beta(\alpha,\beta)}, prior for arm 1 which
stands for the control. Default value is set to 1.}

\item{beta1}{\eqn{\beta} in the \eqn{beta(\alpha,\beta)}, prior for arm 1 which
stands for the control. Default value is set to 1.}

\item{alpha2}{\eqn{\alpha} in the \eqn{beta(\alpha,\beta)}, prior for arm 2.
Default value is set to \code{alpha1}.}

\item{beta2}{\eqn{\beta} in the \eqn{beta(\alpha,\beta)}, prior for arm 2.
Default value is set to \code{beta1}.}

\item{alpha3}{\eqn{\alpha} in the \eqn{beta(\alpha,\beta)} prior for arm 3.
Default value is set to \code{alpha1}.}

\item{beta3}{\eqn{\beta} in the \eqn{beta(\alpha,\beta)}, prior for arm 3.
Default value is set to \code{beta1}.}

\item{alpha4}{\eqn{\alpha} in the \eqn{beta(\alpha,\beta)} prior for arm 4.
Default value is set to \code{alpha1}.}

\item{beta4}{\eqn{\beta} in the \eqn{beta(\alpha,\beta)}, prior for arm 4.
Default value is set to \code{beta1}.}

\item{alpha5}{\eqn{\alpha} in the \eqn{beta(\alpha,\beta)} prior for arm 5.
Default value is set to \code{alpha1}.}

\item{beta5}{\eqn{\beta} in the \eqn{beta(\alpha,\beta)}, prior for arm 5.
Default value is set to \code{beta1}.}

\item{minstart}{a specified number of participants when one starts to check decision rules.}

\item{deltaa}{a vector of minimal effect expected to be observed for early futility stopping in
each arm is approximately \eqn{1\%}. The length of this parameter is \code{armn}-1.}

\item{tpp}{indicator of \code{tp} equals to n/2N. When \code{tp} is n/2N, \code{tpp} should be assigned 1. Default value is set to 0.}

\item{deltaa1}{a vector of pre-specified minimal effect size expected to be observed at the final stage
for each arm. The length of this parameter is \code{armn}-1.}

\item{side}{direction of a one-sided test, with values 'upper' or 'lower'.}

\item{output}{control the output of \code{brar_select_au_binary}. If user does not specify anything, the function returns
the entire data set used to select the stopping boundary for each iteration. If the user specifies 'B', the function
only returns the selected stopping boundary for each iteration.}

\item{...}{additional arguments to be passed to \code{\link[stats]{integrate}} (such as rel.tol) from this function.}
}
\value{
A list of results generated from formula \eqn{Pr(p_k>p_{control}+\delta|data_{t-1})} at each step.
Note that before final stage of the trial, test statistics is calculated from \code{deltaa}, and test statistics is
calculated from \code{deltaa1} at the final stage.
}
\description{
\code{brar_select_au_binary} involves selecting au in Bayesian Response-Adaptive Randomization with a control group
for binary outcomes with two to five arms. The prior distributions follow Beta (\eqn{beta(\alpha,\beta)}) distributions
and can be specified individually for each arm.
}
\details{
This function generates a data set or a value in one iteration for selecting the appropriate au using Bayesian
response-adaptive randomization with a control group under null hypotheses with no delay and delay scenarios.
The function can handle trials with up to 5 arms for binary outcomes. This function uses the formula
\eqn{\frac{Pr(p_k=max\{p_1,...,p_K\})^tp} {\sum_{k=1}^{K}{Pr(p_k=max\{p_1,...,p_K\})^tp}}} with \code{side} equals to 'upper'
and \eqn{\frac{Pr(p_k=min\{p_1,...,p_K\})^tp} {\sum_{k=1}^{K}{Pr(p_k=min\{p_1,...,p_K\})^tp}}} 
with \code{side} equals to 'lower', utilizing available data at each step.
}
\examples{
#brar_select_au_binary with delayed responses follow a normal distribution with
#a mean of 30 days and a standard deviation of 3 days, where h1=c(0.2,0.4), tp=0.5,
#early futility stopping is set at -0.085, and the minimal effect size is 0.1.
set.seed(123)
stopbound1<-lapply(1:100,function(x){brar_select_au_binary(Pats=10,
nMax=50000,TimeToOutcome=expression(rnorm( length( vStartTime ),30, 3)),
enrollrate=0.1,N1=24,armn=2,h=c(0.3,0.3),N2=224,tp=0.5,armlabel=c(1,2),
blocksize=4,alpha1=1,beta1=1,alpha2=1,beta2=1,minstart=24,deltaa=-0.051,
tpp=0,deltaa1=0.1,side='upper')})
simf<-list()
for (xx in 1:100){
    if (any(stopbound1[[xx]][24:223,2]<0.01)){
      simf[[xx]]<-NA
   }  else{
      simf[[xx]]<-stopbound1[[xx]][224,2]
 }
}
simf1<-do.call(rbind,simf)
sum(is.na(simf1))/100  #1, achieve around 1\% futility
sum(simf1>0.436,na.rm=TRUE)/100  #0.05
#the selected stopping boundary is 0.436 with an overall upper one-sided type I
#error of 0.05, based on 100 simulations. It is recommended to conduct more simulations
#to obtain a more accurate au.
}
\references{
<<<<<<< HEAD
\insertRef{Wathen2017}{RARtrials}
=======
\insertRef{Wathen2017}{RAR}
>>>>>>> 61147631f1f0ee0048e12fc72aae70846a4c38ac
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
