% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gamma_to_chisq.r
\name{convert_gamma_to_chisq}
\alias{convert_gamma_to_chisq}
\title{convert_gamma_to_chisq}
\usage{
convert_gamma_to_chisq(gpar)
}
\arguments{
\item{gpar}{a list of parameters including m, V, a, b from a normal-inverse-gamma distribution.}
}
\value{
a list of parameters of a normal-inverse-gamma distribution.
}
\description{
Convert parameters from a normal-inverse-gamma distribution to a
normal-inverse-chi-squared distribution.
}
\details{
This function convert parameters from a normal-inverse-gamma distribution to a
normal-inverse-chi-squared distribution.
}
\examples{
convert_gamma_to_chisq(list(V=1/2,a=0.5,m=9.1/100,b=0.00002))
}
\references{
\insertRef{Kevin2007}{RARtrials}
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
