% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabcd_min_var.r
\name{dabcd_min_var}
\alias{dabcd_min_var}
\title{dabcd_min_var}
\usage{
dabcd_min_var(NN, Ntotal1, armn, type, dabcd = 0, gamma = 2)
}
\arguments{
\item{NN}{a vector representing the number of participants with success results for each arm
estimated from the current data .}

\item{Ntotal1}{a vector representing the total number of participants for each arm
estimated from the current data.}

\item{armn}{number of total arms in the trial.}

\item{type}{allocation type, with choices from 'RSIHR' and 'Neyman'.}

\item{dabcd}{an indicator of whether to apply Hu & Zhang's formula (\insertCite{Hu2004}{RAR}), with choices from 0 and 1.
1 represents allocation probabilities calculated using Hu & Zhang's formula;
0 represents allocation probabilities calculated before applying Hu & Zhang's formula.
Default value is set to 0.}

\item{gamma}{tuning parameter in Hu & Zhang's formula (\insertCite{Hu2004}{RAR}). When \code{dabcd}=0, this parameter does not need
to be specified. Default value is set to 2.}
}
\value{
A vector of allocation probabilities to each arm.
}
\description{
\code{dabcd_min_var} can be used for doubly adaptive biased coin design with minimal variance
strategy for binary outcomes, targeting generalized Neyman allocation and generalized RSIHR allocation. The return 
of this function is a vector of allocation probabilities to each arm, with the pre-specified number of participants in the trial.
}
\details{
The function simulates allocation probabilities for doubly adaptive biased coin design with minimal variance strategy targeting
generalized Neyman allocation and generalized RSIHR allocation with 2-5 arms. The output of this function is based on Hu \& Zhang's formula \insertCite{Hu2004}{RAR}.
With more than two armd the one-sided nominal level of each test is \code{alphaa} divided by \code{arm*(arm-1)/2}; a Bonferroni correction.
}
\examples{
dabcd_min_var(NN=c(54,67,85,63,70),Ntotal1=c(100,88,90,94,102),armn=5, type='Neyman')
dabcd_min_var(NN=c(54,67,85,63),Ntotal1=c(100,88,90,94),armn=4,type='RSIHR')
}
\references{
\insertRef{Hu2004}{RAR}
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
