% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flgi_stop_bound_flgi_unk_var.r
\name{flgi_stop_bound_flgi_unk_var}
\alias{flgi_stop_bound_flgi_unk_var}
\title{flgi_stop_bound_flgi_unk_var}
\usage{
flgi_stop_bound_flgi_unk_var(
  Gittinstype,
  df,
  gittins = NULL,
  Pats,
  nMax,
  TimeToOutcome,
  enrollrate,
  K,
  noRuns2,
  Tsize,
  block,
  rule,
  prior_n,
  prior_mean1,
  prior_sd1,
  mean,
  sd,
  side
)
}
\arguments{
\item{Gittinstype}{type of Gittins indices, should be set to 'UNKV' in this function}

\item{df}{discount factor which is the multiplier for loss at each additional patient in the future.
Available values are 0.5, 0.6, 0.7, 0.8, 0.9, 0.95, 0.99 and 0.995. The maximal sample size can be up to 10000.}

\item{gittins}{user specified Gittins indices for calculation in this function. If \code{gittins} is provided,
\code{Gittinstype} and \code{df} should be NULL.}

\item{Pats}{the number of patients accrued within a certain time frame indicates the
count of individuals who have been affected by the disease during that specific period,
for example, a month or a day. If this number is 10, it represents that
10 people have got the disease within the specified time frame.}

\item{nMax}{the maximal accrued number of patients with the disease, this number
should be chosen carefully to ensure a sufficient number of patients are simulated,
especially when considering the delay mechanism.}

\item{TimeToOutcome}{Representation of the time distribution of delayed responses. The accrual times
could be a month, a week or any other time frame. When the unit changes,
the number of TimeToOutcome should also change. It can be in the format
of expression(rnorm( length( vStartTime ),30, 3)), representing delayed responses
with a normal distribution, where the mean is 30 days and the standard deviation is 3 days.
These related functions are adapted from \url{http://github.com/kwathen/IntroBayesianSimulation}.
Refer to the website for more details.}

\item{enrollrate}{probability that patients in the population can enroll in the trial.
This parameter is related to the number of people who have been affected by the disease in the population,
following an exponential distribution.}

\item{K}{number of total arms in the trial.}

\item{noRuns2}{number of simulations for simulated allocation probabilities within each block. Default value is
set to 100 times, which is recommended in Villar et al., 2015.}

\item{Tsize}{maximal sample size for the trial.}

\item{block}{block size.}

\item{rule}{rules can be used in this function, with values 'FLGI PM', 'FLGI PD' or 'CFLGI'.
'FLGI PM' stands for making decision based on posterior mean;
'FLGI PD' stands for making decision based on posterior distribution;
'CFLGI' stands for controlled forward-looking Gittins index.}

\item{prior_n}{a vector representing the number of observations assumed in prior distribution, eg: c(1,1) for a two-armed trial.}

\item{prior_mean1}{a vector representing mean of observations assumed in prior distributions, eg: c(0,0,0) for a three-armed trial,
rep(0,K) can be used to simplify the process. If a negative effect is expected, adjust the mean to a negative value.}

\item{prior_sd1}{a vector representing the standard deviation of observations assumed in prior distribution, eg: rep(1,3) for a three-armed trial.}

\item{mean}{a vector of mean hypotheses, for example, as c(0.1,0.1) where 0.1 stands for the mean
for both groups. Another example is c(0.1,0.3) where 0.1 and 0.3 stand for the mean for the control and
a treatment group, respectively.}

\item{sd}{a vector of standard deviation hypotheses, for example, as c(0.64,0.64) where 0.64 stands for the standard deviation
for both groups. Another example is c(0.64,0.4) where 0.64 and 0.4 stand for the standard deviation for the control and
a treatment group, respectively.}

\item{side}{direction of one-sided test with the values of 'upper' or 'lower'.}
}
\value{
Value of T test statistics for one trial.
}
\description{
Function for simulating cut-off values at the final stage using the forward-looking Gittins index
and the controlled forward-looking Gittins index algorithm for continuous outcomes with known variance in trials with
2-5 arms. The prior distributions follow Normal-Inverse-Gamma (NIG) (\eqn{NIG(V,m,a,b)}) distributions and should be the same for each arm.
}
\details{
This function simulates trials using the forward-looking Gittins index and the
controlled forward-looking Gittins index algorithm under both no delay and delay scenarios to obtain
cut-off values at the final stage, with control of type I error. The user is expected to run this function
multiple times to determine a reasonable cut-off value for statistical inference.
}
\examples{
#forward-looking Gittins index with delayed responses follow a normal distribution
#with a mean of 30 days and a standard deviation of 3 days
set.seed(12345)
stopbound1<-lapply(1:100,function(x){
flgi_stop_bound_flgi_unk_var(Gittinstype='UNKV',df=0.995,Pats=10,nMax=50000,
TimeToOutcome=expression(rnorm( length( vStartTime ),30, 3)),enrollrate=0.5,
K=3,noRuns2=100,Tsize=120,block=8,rule='FLGI PM',prior_n=rep(2,3),
prior_mean1=rep(0,3),prior_sd1=rep(1,3),mean=c(-0.05,-0.05,-0.05),
sd=c(0.346,0.346,0.346),side='upper')})
stopbound1a<-do.call(rbind,stopbound1)
sum(stopbound1a[,1]>1.885| stopbound1a[,2]>1.885 )/100 #0.05
#the selected cut-off value is 1.885 with an overall upper one-sided type I
#error of 0.05, based on 100 simulations.
#It is recommended to conduct more simulations to obtain a more accurate cut-off value.

#forward-looking Gittins index with no delay responses
set.seed(12345)
stopbound2<-lapply(1:100,function(x){
flgi_stop_bound_flgi_unk_var(Gittinstype='UNKV',df=0.995,Pats=10,nMax=50000,
TimeToOutcome=0,enrollrate=0.5,K=3,noRuns2=100,Tsize=120,block=8,
rule='FLGI PM',prior_n=rep(2,3),prior_mean1=rep(0,3),prior_sd1=rep(1,3),
mean=c(0.05,0.05,0.05),sd=c(0.346,0.346,0.346),side='lower')})
stopbound2a<-do.call(rbind,stopbound2)
sum(stopbound2a[,1]<=-2.0 | stopbound2a[,2]<=-2.0 )/100 #0.05
#the selected cut-off value is -2.0 with an overall lower one-sided type I
#error of 0.05, based on 100 simulations.
#It is recommended to conduct more simulations to obtain a more accurate cut-off value.

}
\references{
\insertRef{Williamson2019}{RARtrials}
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
