% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgreater.r
\name{pgreater}
\alias{pgreater}
\title{pgreater}
\usage{
pgreater(a, b, A, B, delta, side, ...)
}
\arguments{
\item{a}{\eqn{\alpha} in \eqn{beta(\alpha,\beta)}, current accumulated \eqn{\alpha} for the treatment group which is compared to the control group.}

\item{b}{\eqn{\beta} in \eqn{beta(\alpha,\beta)}, current accumulated \eqn{\beta} for the treatment group which is compared to the control group.}

\item{A}{\eqn{\alpha} in \eqn{beta(\alpha,\beta)}, current accumulated \eqn{\alpha} for the control group.}

\item{B}{\eqn{\beta} in \eqn{beta(\alpha,\beta)}, current accumulated \eqn{\beta} for the control group.}

\item{delta}{expected difference in success probabilities between the control group and the treatment group.}

\item{side}{direction of a one-sided test, with values 'upper' or 'lower'.}

\item{...}{additional arguments to be passed to stats::integrate() (such as rel.tol) from this function.}
}
\value{
a posterior probability of \eqn{Pr(p_k>p_{control}+\delta|data)} with \code{side} equals to 'upper';
a posterior probability of \eqn{Pr(p_{control}>p_k+\delta|data)} with \code{side} equals to 'lower'.
}
\description{
Calculate the futility stopping probability in Bayesian response-adaptive randomization with
a control group using Thall \& Wathen method for binary outcomes. The prior distributions follow
Beta (\eqn{beta(\alpha,\beta)}) distributions and can be specified individually for each treatment group.
}
\details{
This function calculates the results of \eqn{Pr(p_k>p_{control}+\delta|data)} for \code{side} equals to
'upper' and the results of \eqn{Pr(p_{control}>p_k+\delta|data)} for \code{side} equals to 'lower'.
The result indicates the posterior probability of stopping a treatment group due to futility around \eqn{1\%} in Bayesian
response-adaptive randomization with a control arm using Thall \& Wathen method, with accumulated results
during the conduct of trials.
}
\examples{
pgreater(a=5, b=19, A=8, B=10, delta=0.1, side='upper')
pgreater(a=58, b=68, A=65, B=79, delta=0, side='lower')
}
\references{
<<<<<<< HEAD
\insertRef{Wathen2017}{RARtrials}
=======
\insertRef{Wathen2017}{RAR}
>>>>>>> 61147631f1f0ee0048e12fc72aae70846a4c38ac
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
