% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmax.r
\name{pmax}
\alias{pmax}
\title{pmax}
\usage{
pmax(
  outcome,
  armn,
  a1 = NULL,
  b1 = NULL,
  a2 = NULL,
  b2 = NULL,
  a3 = NULL,
  b3 = NULL,
  a4 = NULL,
  b4 = NULL,
  a5 = NULL,
  b5 = NULL,
  mean1 = NULL,
  sd1 = NULL,
  mean2 = NULL,
  sd2 = NULL,
  mean3 = NULL,
  sd3 = NULL,
  mean4 = NULL,
  sd4 = NULL,
  mean5 = NULL,
  sd5 = NULL,
  par1 = NULL,
  par2 = NULL,
  par3 = NULL,
  par4 = NULL,
  par5 = NULL,
  side,
  ...
)
}
\arguments{
\item{outcome}{type of outcomes, with values of choices from 'binary', 'UNKV' and 'KV', representing binary outcomes,
continuous outcomes with known variance and continuous outcomes with unknown variance respectively.
When 'binary' is specified, \code{mean1} to \code{mean5}, \code{sd1} to \code{sd5}, \code{par1} to \code{par5} should be NULL.
When 'UNKV' is specified, \code{a1} to \code{a5}, \code{b1} to \code{b5}, \code{mean1} to \code{mean5}, \code{sd1} to \code{sd5} should be NULL.
When 'KV' is specified, \code{a1} to \code{a5}, \code{b1} to \code{b5}, \code{par1} to \code{par5} should be NULL.}

\item{armn}{number of arms in the trial with values up to 5. When \code{armn}=2,
only \code{a1} to \code{a2} and \code{b1} to \code{b2} OR \code{mean1} to \code{mean2} and \code{sd1} to \code{sd2} 
OR \code{par1} to \code{par2} need to be specified.
When \code{armn}=3, only \code{a1} to \code{a3} and  \code{b1} to \code{b3} OR \code{mean1} to \code{mean3} and \code{sd1} to \code{sd3} 
OR \code{par1} to \code{par3} need to be specified.
When \code{armn}=4, only \code{a1} to \code{a4} and  \code{b1} to \code{b4} OR \code{mean1} to \code{mean4} and \code{sd1} to \code{sd4} 
OR \code{par1} to \code{par4} need to be specified.
When \code{armn}=5, \code{a1} to \code{a5} and  \code{b1} to \code{b5} OR \code{mean1} to \code{mean5} and \code{sd1} to \code{sd5} 
OR \code{par1} to \code{par5} need to be specified.}

\item{a1}{\eqn{\alpha} in prior \eqn{beta(\alpha,\beta)} for the arm to calculate the allocation probability of.}

\item{b1}{\eqn{\beta} in prior \eqn{beta(\alpha,\beta)} for the arm to calculate the allocation probability of.}

\item{a2}{\eqn{\alpha} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{b2}{\eqn{\beta} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{a3}{\eqn{\alpha} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{b3}{\eqn{\beta} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{a4}{\eqn{\alpha} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{b4}{\eqn{\beta} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{a5}{\eqn{\alpha} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{b5}{\eqn{\beta} in prior \eqn{beta(\alpha,\beta)} for one of the remaining arms.}

\item{mean1}{mean in prior Normal(mean,sd) for the arm to calculate the allocation probability of.}

\item{sd1}{sd in prior Normal(mean,sd) for the arm to calculate the allocation probability of.}

\item{mean2}{mean in prior Normal(mean,sd) for one of the remaining arms.}

\item{sd2}{sd in prior Normal(mean,sd) for one of the remaining arms.}

\item{mean3}{mean in prior Normal(mean,sd) for one of the remaining arms.}

\item{sd3}{sd in prior Normal(mean,sd) for one of the remaining arms.}

\item{mean4}{mean in prior Normal(mean,sd) for one of the remaining arms.}

\item{sd4}{sd in prior Normal(mean,sd) for one of the remaining arms.}

\item{mean5}{mean in prior Normal(mean,sd) for one of the remaining arms.}

\item{sd5}{sd in prior Normal(mean,sd) for one of the remaining arms.}

\item{par1}{a vector of parameters for the arm with a Normal-Inverse-Chi-Squared prior
to calculate the allocation probability of.}

\item{par2}{a vector of parameters for one of the remaining arms with a Normal-Inverse-Chi-Squared prior.}

\item{par3}{a vector of parameters for one of the remaining arms with a Normal-Inverse-Chi-Squared prior.}

\item{par4}{a vector of parameters for one of the remaining arms with a Normal-Inverse-Chi-Squared prior.}

\item{par5}{a vector of parameters for one of the remaining arms with a Normal-Inverse-Chi-Squared prior.}

\item{side}{direction of a one-sided test, with values 'upper' or 'lower'.}

\item{...}{additional arguments to be passed to \code{\link[stats]{integrate}} (such as rel.tol) from this function.}
}
\value{
a probability that a particular arm is the best in trials up to five arms.
}
\description{
Calculate probability that a particular arm is the best in a trial using Bayesian response-adaptive randomization with
a control group using Thall \& Wathen method. The prior distributions follow Beta (\eqn{beta(\alpha,\beta)}) distributions
for binary outcomes, Normal (\eqn{N(mean,sd)}) distributions for continuous outcomes with known variance, and 
Normal-Inverse-Gamma (NIG) (\eqn{NIG(V,m,a,b)}) distributions which are equivalent to Normal-Inverse-Chi-Squared distributions 
for continuous outcomes with unknown variances for each arm and can be specified individually.
}
\details{
This function calculates the results of formula \eqn{Pr(p_k=max\{p_1,...,p_K\})} for
\code{side} equals to 'upper' and the results of formula \eqn{Pr(p_k=min\{p_1,...,p_K\})} for
\code{side} equals to 'lower'. This function returns the probability that the posterior probability of arm
\eqn{k} is maximal or minimal in trials with up to five arms.
}
\examples{
pmax(outcome='binary',armn=5,a1=8,b1=10,a2=5,b2=19,a3=8,b3=21,
a4=6, b4=35, a5=15, b5=4, side='upper')
pmax(outcome='KV',armn=4,mean1=8, sd1=2,mean2=5,sd2=1,mean3=8,
sd3=1.5,mean4=6,sd4=2,side='upper')
para<-list(V=1/2,a=0.5,m=9.1/100,b=0.00002)
par<-convert_gamma_to_chisq(para)
set.seed(123451)
y1<-rnorm(100,0.091,0.009)
par11<-update_par_nichisq(y1, par)
set.seed(123452)
y2<-rnorm(90,0.09,0.009)
par22<-update_par_nichisq(y2, par)
set.seed(123453)
y3<-rnorm(110,0.0892,0.009)
par33<-update_par_nichisq(y3, par)
pmax(outcome='UNKV',armn=3,par1=par11,par2=par22,par3=par33,side='upper')
pmax(outcome='UNKV',armn=3,par1=par11,par2=par22,par3=par33,side='lower')
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
