% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR-package.r
\docType{package}
\name{RARtrials-package}
\alias{RARtrials}
\alias{RARtrials-package}
\title{RARtrials: Response-Adaptive Randomisation in Clinical Trials}
\description{
RAR package is designed for simulating most popular response-adaptive randomization 
methods in the literature with comparisons of each treatment group to a control 
group under no delay and delayed scenarios. All the designs are based on one-sided 
tests with a choice from values of 'upper' and 'lower'. The general assumption is 
that binary outcomes follow a Binomial distribution, while continuous outcomes follow 
a normal distribution. Additionally, the number of patients accrued in the population 
follows a Poisson distribution and users can specify the enrollment rate of patients 
enrolled in the trial.
}
\author{
\strong{Maintainer}: Chuyao Xu \email{cxu870@aucklanduni.ac.nz}

Authors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
  \item Alain Vandal \email{alain.vandal@auckland.ac.nz} [thesis advisor]
}

Other contributors:
\itemize{
  \item Sofia S. Villar \email{sofia.villar@mrc-bsu.cam.ac.uk} [copyright holder]
  \item Kyle J. Wathen \email{kylewathen@gmail.com} [copyright holder]
}

}
\keyword{internal}
