% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_A_optimal_unknown_var.r
\name{sim_A_optimal_unknown_var}
\alias{sim_A_optimal_unknown_var}
\title{sim_A_optimal_unknown_var}
\usage{
sim_A_optimal_unknown_var(
  Pats,
  nMax,
  TimeToOutcome,
  enrollrate,
  N1,
  N2,
  armn,
  mean,
  sd,
  alphaa = 0.025,
  armlabel,
  side
)
}
\arguments{
\item{Pats}{the number of patients accrued within a certain time frame indicates the
count of individuals who have been affected by the disease during that specific period,
for example, a month or a day. If this number is 10, it represents that
10 people have got the disease within the specified time frame.}

\item{nMax}{the maximal accrued number of patients with the disease, this number
should be chosen carefully to ensure a sufficient number of patients are simulated,
especially when considering the delay mechanism.}

\item{TimeToOutcome}{Representation of the time distribution of delayed responses. The accrual times
could be a month, a week or any other time frame. When the unit changes,
the number of TimeToOutcome should also change. It can be in the format
of expression(rnorm( length( vStartTime ),30, 3)), representing delayed responses
with a normal distribution, where the mean is 30 days and the standard deviation is 3 days.
These related functions are adapted from \url{http://github.com/kwathen/IntroBayesianSimulation}.
Refer to the website for more details.}

\item{enrollrate}{probability that patients in the population can enroll in the trial.
This parameter is related to the number of people who have been affected by the disease in the population,
following an exponential distribution.}

\item{N1}{number of participants with equal randomization in the 'initialization' period.
Recommend using 10 percent of the total sample size.}

\item{N2}{maximal sample size for the trial.}

\item{armn}{number of total arms in the trial.}

\item{mean}{a vector of hypotheses of mean for all arms in the trial,
with the first one serving as the control group.}

\item{sd}{a vector of hypotheses of standard deviation for all arms in the trial,
with the first one serving as the control group.}

\item{alphaa}{the overall type I error to be controlled for the one-sided test. Default value is set to 0.025.}

\item{armlabel}{a vector of arm labels with an example of c(1, 2), where 1 and 2 describes
how each arm is labeled in a two-armed trial.}

\item{side}{direction of a one-sided test, with values 'upper' or 'lower'.}
}
\value{
A list of results, including final decision based on the T test statistics with 1 stands
for effectiveness and 0 stands for not selected, T test statistics and the simulated data set for one trial.
The simulated data set includes 5 columns: participant ID number, enrollment time, observed time of results,
allocated arm, and participants' results.
In the final decision, 1 refers to selected, and 0 stands for not selected.
}
\description{
\code{sim_A_optimal_unknown_var} simulates a trial for continuous endpoints with unknown variances,
and the allocation probabilities change based on results of accumulated participants in the trial.
}
\details{
This function aims to minimize the criteria \eqn{tr[M^{-1}(\mathbf{\rho})]}
and to minimize the overall variance of pairwise comparisons. It is analogous to Neyman
allocation, specifically designed for continuous endpoints with known variances.
With more than two arms the one-sided nominal level of each test is \code{alphaa}
divided by \code{arm*(arm-1)/2}; a Bonferroni correction.
}
\examples{
#Run the function with delayed responses follow a normal distribution with
#a mean of 30 days and a standard deviation of 3 days under null hypothesis
#in a three-armed trial
sim_A_optimal_unknown_var(Pats=10,nMax=50000,TimeToOutcome=expression(
rnorm( length( vStartTime ),30, 3)),enrollrate=0.1,N1=12,N2=132,armn=3,
mean=c(9.1/100,9.1/100,9.1/100),sd=c(0.009,0.009,0.009),alphaa=0.025,
armlabel = c(1,2,3),side='upper')

#Run the function with delayed responses follow a normal distribution with
#a mean of 30 days and a standard deviation of 3 days under alternative hypothesis
#in a three-armed trial
sim_A_optimal_unknown_var(Pats=10,nMax=50000,TimeToOutcome=expression(
rnorm( length( vStartTime ),30, 3)),enrollrate=0.1,N1=12,N2=132,armn=3,
mean=c(9.1/100,9.28/100,9.28/100),sd=c(0.009,0.009,0.009),alphaa=0.025,
armlabel = c(1,2,3),side='upper')
}
\references{
\insertRef{Oleksandr2013}{RARtrials}
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
