% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_par_nichisq.r
\name{update_par_nichisq}
\alias{update_par_nichisq}
\title{update_par_nichisq}
\usage{
update_par_nichisq(y, par)
}
\arguments{
\item{y}{observed data.}

\item{par}{a vector of current parameters from a Normal-Inverse-Chi-Squared distribution.}
}
\value{
a list of parameters for a posterior Normal-Inverse-Chi-Squared distribution incorporating available data.
}
\description{
Update hyperparameters of a Normal-Inverse-Chi-Squared distribution
}
\details{
This function updates hyperparameters of a Normal-Inverse-Chi-Squared distribution
with available data to hyperparameters of a posterior Normal-Inverse-Chi-Squared distribution.
Those updated hyperparameters can be converted to hyperparameters in a Normal-Inverse-Gamma distribution
for continuous outcomes with unknown variances.
}
\examples{
para<-list(V=1/2,a=0.5,m=9.1/100,b=0.00002)
par<-convert_gamma_to_chisq(para)
set.seed(123451)
y1<-rnorm(100,0.091,0.009)
update_par_nichisq(y1, par)
set.seed(123452)
y2<-rnorm(90,0.09,0.009)
update_par_nichisq(y2, par)
}
\references{
\insertRef{Kevin2007}{RAR}
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
