% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gamma_to_chisq.r
\name{convert_gamma_to_chisq}
\alias{convert_gamma_to_chisq}
\title{convert_gamma_to_chisq}
\usage{
convert_gamma_to_chisq(gpar)
}
\arguments{
\item{gpar}{a list of parameters including m, V, a, b from a normal-inverse-gamma distribution.}
}
\value{
a list of parameters of a normal-inverse-gamma distribution.
}
\description{
Convert parameters from a normal-inverse-gamma distribution to a
normal-inverse-chi-squared distribution.
}
\details{
This function convert parameters from a normal-inverse-gamma distribution to a
normal-inverse-chi-squared distribution.
}
\examples{
convert_gamma_to_chisq(list(V=1/2,a=0.5,m=9.1/100,b=0.00002))
}
\references{
{Wathen, J. K., & Thall, P. F. (2017). A simulation study of outcome adaptive
randomization in multi-arm clinical trials. Clinical trials (London, England), 14(5), 432–440.
https://doi.org/10.1177/1740774517692302}

\insertRef{Kevin2007}{RAR}
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal
}
