% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flgi_cut_off_binary.r
\name{flgi_cut_off_binary}
\alias{flgi_cut_off_binary}
\title{Cut-off Value of Forward-looking Gittins index rule in Binary Endpoint}
\usage{
flgi_cut_off_binary(
  Gittinstype,
  df,
  gittins = NULL,
  Pats,
  nMax,
  TimeToOutcome,
  enrollrate,
  I0,
  K,
  noRuns2 = 100,
  Tsize,
  ptrue,
  block,
  rule,
  ztype
)
}
\arguments{
\item{Gittinstype}{type of Gittins indices, should be set to 'binary' in this function.}

\item{df}{discount factor which is the multiplier for loss at each additional patient in the future.
Available values are 0, 0.5, 0.7, 0.99 and 0.995. The maximal sample size can be up to 2000.}

\item{gittins}{user specified Gittins indices for calculation in this function. Recommend using the
\code{bmab_gi_multiple_ab} function from \code{gittins} package. If \code{gittins} is provided,
\code{Gittinstype} and \code{df} should be NULL.}

\item{Pats}{the number of patients accrued within a certain time frame indicates the
count of individuals who have been affected by the disease during that specific period,
for example, a month or a day. If this number is 10, it represents that
10 people have got the disease within the specified time frame.}

\item{nMax}{the maximal accrued number of patients with the disease, this number
should be chosen carefully to ensure a sufficient number of patients are simulated,
especially when considering the delay mechanism.}

\item{TimeToOutcome}{Representation of the time distribution of delayed responses. The accrual times
could be a month, a week or any other time frame. When the unit changes,
the number of TimeToOutcome should also change. It can be in the format
of expression(rnorm( length( vStartTime ),30, 3)), representing delayed responses
with a normal distribution, where the mean is 30 days and the standard deviation is 3 days.
These related functions are adapted from \url{https://github.com/kwathen/IntroBayesianSimulation}.
Refer to the website for more details.}

\item{enrollrate}{probability that patients in the population can enroll in the trial.
This parameter is related to the number of people who have been affected by the disease in the population,
following an exponential distribution.}

\item{I0}{a matrix with K rows and 2 columns, where the numbers inside are equal to the prior parameters, and
K is equal to the total number of arms. For example, matrix(1,nrow=2,ncol=2) means that the prior
distributions for two-armed trials are beta(1,1); matrix(c(2,3),nrow=2,ncol=2,byrow = TRUE) means that the prior
distributions for two-armed trials are beta(2,3). The first column represents the prior of the number of successes,
and the second column represents the prior of the number of failures.}

\item{K}{number of total arms in the trial.}

\item{noRuns2}{number of simulations for simulated allocation probabilities within each block. Default value is
set to 100, which is recommended in \insertCite{Villar2015}{RARtrials}.}

\item{Tsize}{maximal sample size for the trial.}

\item{ptrue}{a vector of hypotheses, for example, as c(0.1,0.1) where 0.1 stands for the success probability
for both groups. Another example is c(0.1,0.3) where 0.1 and 0.3 stand for the success probability for the control and
the treatment group, respectively.}

\item{block}{block size.}

\item{rule}{rules can be used in this function, with values 'FLGI PM', 'FLGI PD' or 'CFLGI'.
'FLGI PM' stands for making decision based on posterior mean;
'FLGI PD' stands for making decision based on posterior distribution;
'CFLGI' stands for controlled forward-looking Gittins index.}

\item{ztype}{Z test statistics, with choice of values from 'pooled' and 'unpooled'.}
}
\value{
Value of Z test statistics for one trial.
}
\description{
Function for simulating cut-off values at the final stage using the forward-looking Gittins index
and the controlled forward-looking Gittins index algorithm for binary outcomes in trials with 2-5 arms.
The conjugate prior distributions follow Beta (\eqn{Beta(\alpha,\beta)}) distributions and should be the same for each
arm.
}
\details{
This function simulates trials using the forward-looking Gittins index and the
controlled forward-looking Gittins index algorithm under both no delay and delay scenarios to obtain
cut-off values at the final stage, with control of type I error. The user is expected to run this function
multiple times to determine a reasonable cut-off value for statistical inference.
}
\examples{
#forward-looking Gittins index rule with delayed responses follow a normal distribution
#with a mean of 30 days and a standard deviation of 3 days
set.seed(12345)
stopbound1<-lapply(1:10,function(x){ flgi_cut_off_binary(Gittinstype='Binary',df=0.5,Pats=10,
nMax=50000,TimeToOutcome=expression(rnorm( length( vStartTime ),30, 3)),
enrollrate=0.1,I0= matrix(1,nrow=2,ncol=2),K=2,Tsize=256,ptrue=c(0.2,0.2),block=2,
rule='FLGI PM',ztype='unpooled')})
stopbound1a<-do.call(rbind,stopbound1)
sum(stopbound1a< (-1.82) )/10 #0.1
#the selected cut-off value is -1.82 with an overall lower one-sided type I
#error of 0.1, based on 10 simulations.
#It is recommended to conduct more simulations (i.e.,1000) to obtain an accurate cut-off value.

#forward-looking Gittins index rule with delayed responses follow a normal distribution
#with a mean of 30 days and a standard deviation of 3 days
set.seed(12345)
stopbound2<-lapply(1:10,function(x){ flgi_cut_off_binary(Gittinstype='Binary',df=0,Pats=10,
nMax=50000,TimeToOutcome=expression(rnorm( length( vStartTime ),30, 3)),
enrollrate=0.1,I0= matrix(1,nrow=2,ncol=2),K=2,Tsize=256,ptrue=c(0.2,0.2),block=2,
rule='FLGI PM',ztype='unpooled')})
stopbound2a<-do.call(rbind,stopbound2)
sum(stopbound2a>0.5)/10 #0.1
#the selected cut-off value is 0.5 with an overall upper one-sided type I
#error of 0.1, based on 10 simulations.
#It is recommended to conduct more simulations (i.e.,1000) to obtain an accurate cut-off value.

}
\references{
\insertRef{Gittins2011}{RARtrials}
\insertRef{Villar2015}{RARtrials}
}
\author{
Chuyao Xu, Thomas Lumley, Alain Vandal, Villar Sofia S.
}
